requirement.txt
fastapi==0.95.2
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
